backend/pptagent/requirements.txt
fastapi
uvicorn
google-genai
google-adk==1.20.0
httpx
a2a-sdk==0.2.10
litellm
python-dotenv
fastmcp==2.2.5
asyncclick
pytest
click
BeautifulSoup4
lxml
elasticsearch==7.12.1
numpy<2.0.0
pika
arxiv
gunicorn